Fixed bug where Audio does not stop playing when TimeUp Modal is closed.
